backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
langserve>=0.0.51
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langgraph>=0.0.20
ollama>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sse-starlette>=1.6.0
httpx>=0.26.0
